alteração na lista de favoritos
